Added new features

+++ src/main/java/com/olympus/system/hawkdeskpos/dto/StockBatchDto.java
@@ -8,7 +8,7 @@ public record StockBatchDto(
         String itemSku,      // item-level SKU
         String displaySku,   // variant SKU if set, otherwise item SKU
         String batch,        // batch number / label
-        int qty,
+        double qty,
         int minLevel,
         double costPrice,
         double sellingPrice,
@@ -16,7 +16,7 @@ public record StockBatchDto(
         java.util.Date expiryDate   // null if no expiry set
 ) {
     public String stockStatus() {
-        if (qty <= 0) return "OUT";
+        if (qty <= 0.0) return "OUT";
         if (qty <= minLevel) return "LOW";
         return "OK";
     }
